Move FetchPreviousTransactions service
